replaced StatHandler with GameStats
