+++ src/main/java/com/randude14/hungergames/listeners/EntityListener.java
@@ -10,8 +10,10 @@ import com.randude14.hungergames.stats.PlayerStat;
 import com.randude14.hungergames.stats.PlayerStat.PlayerState;
 import com.randude14.hungergames.stats.PlayerStat.Team;
 import com.randude14.hungergames.utils.ChatUtils;
+import org.bukkit.entity.Entity;
 
 import org.bukkit.entity.Player;
+import org.bukkit.entity.Projectile;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
@@ -35,20 +37,25 @@ public class EntityListener implements Listener{
 				EntityDamageByEntityEvent newEvent = (EntityDamageByEntityEvent) event;
 				double period = Config.GRACE_PERIOD.getDouble(hurtGame.getSetup());
 				long startTime = hurtGame.getInitialStartTime();
+				Entity damager = newEvent.getDamager() instanceof Projectile ? ((Projectile) newEvent.getDamager()).getShooter() : newEvent.getDamager();
 				if (((System.currentTimeMillis() - startTime) / 1000) < period) {
 					event.setCancelled(true);
-					if (newEvent.getDamager() instanceof Player) {
-						ChatUtils.error((Player) newEvent.getDamager(), "You can't hurt that player during the grace-period!");
+					if (damager instanceof Player) {
+						ChatUtils.error((Player) damager, "You can't hurt that player during the grace-period!");
 					}
 				}
-				if (!Config.TEAMS_ALLOW_FRIENDLY_DAMAGE.getBoolean(hurtGame.getSetup())) {
-					Team hurtTeam = hurtGame.getPlayerStat(player).getTeam();
-					if (newEvent.getDamager() instanceof Player && hurtTeam != null && hurtGame.contains((Player) newEvent.getDamager())){
-						Team hurterTeam = hurtGame.getPlayerStat((Player) newEvent.getDamager()).getTeam();
-						if (hurterTeam != null) {
-							if (hurtTeam.getName().equals(hurterTeam.getName())) {
-								event.setCancelled(true);
-								ChatUtils.error((Player) newEvent.getDamager(), "You can't hurt a player on your team!");
+				if (damager instanceof Player && hurtGame.contains((Player) damager)) {
+					if (!Config.TEAMS_ALLOW_FRIENDLY_DAMAGE.getBoolean(hurtGame.getSetup())) {
+						Team hurtTeam = hurtGame.getPlayerStat(player).getTeam();
+						if (hurtGame.getPlayerStat(player).getState() == PlayerState.WAITING) {
+							event.setCancelled(true);
+						} else if (hurtTeam != null){
+							Team hurterTeam = hurtGame.getPlayerStat((Player) damager).getTeam();
+							if (hurterTeam != null) {
+								if (hurtTeam.getName().equals(hurterTeam.getName())) {
+									event.setCancelled(true);
+									ChatUtils.error((Player) damager, "You can't hurt a player on your team!");
+								}
 							}
 						}
 					}
@@ -70,7 +77,7 @@ public class EntityListener implements Listener{
 		// Games
 		Game game = HungerGames.getInstance().getGameManager().getRawSession(player);
 		if (game != null) {
-			if (game.getState() == GameState.STOPPED) {
+			if (game.getPlayerStat(player).getState() == PlayerState.WAITING) {
 				if (!Config.STOP_TARGETTING.getBoolean(game.getSetup())) return;
 				PlayerStat stat = game.getPlayerStat(player);
 				if (stat != null && stat.getState().equals(PlayerState.WAITING)) event.setCancelled(true);
